+++ tests/test_sensor.py
@@ -19,6 +19,7 @@ from homeassistant.const import (
     STATE_UNKNOWN,
     TEMP_CELSIUS,
     PRESSURE_PSI,
+
 )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
@@ -394,6 +395,7 @@ async def test_tpms_pressure_sensor(hass: HomeAssistant) -> None:
     assert state_fl.attributes.get(ATTR_STATE_CLASS) == SensorStateClass.MEASUREMENT
     assert state_fl.attributes.get(ATTR_UNIT_OF_MEASUREMENT) == PRESSURE_PSI
 
+
     assert (
         state_fl.attributes.get("tpms_last_seen_pressure_timestamp")
         == car_mock_data.VEHICLE_DATA["vehicle_state"][
@@ -409,6 +411,7 @@ async def test_tpms_pressure_sensor(hass: HomeAssistant) -> None:
     assert state_fr.attributes.get(ATTR_DEVICE_CLASS) == SensorDeviceClass.PRESSURE
     assert state_fr.attributes.get(ATTR_STATE_CLASS) == SensorStateClass.MEASUREMENT
     assert state_fr.attributes.get(ATTR_UNIT_OF_MEASUREMENT) == PRESSURE_PSI
+
 
     assert (
         state_fr.attributes.get("tpms_last_seen_pressure_timestamp")
@@ -426,6 +429,7 @@ async def test_tpms_pressure_sensor(hass: HomeAssistant) -> None:
     assert state_rl.attributes.get(ATTR_STATE_CLASS) == SensorStateClass.MEASUREMENT
     assert state_rl.attributes.get(ATTR_UNIT_OF_MEASUREMENT) == PRESSURE_PSI
 
+
     assert (
         state_rl.attributes.get("tpms_last_seen_pressure_timestamp")
         == car_mock_data.VEHICLE_DATA["vehicle_state"][
@@ -441,6 +445,7 @@ async def test_tpms_pressure_sensor(hass: HomeAssistant) -> None:
     assert state_rr.attributes.get(ATTR_DEVICE_CLASS) == SensorDeviceClass.PRESSURE
     assert state_rr.attributes.get(ATTR_STATE_CLASS) == SensorStateClass.MEASUREMENT
     assert state_rr.attributes.get(ATTR_UNIT_OF_MEASUREMENT) == PRESSURE_PSI
+
 
     assert (
         state_rr.attributes.get("tpms_last_seen_pressure_timestamp")